requirements.txt
Django==5.1.5
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.1
Pillow==11.0.0
qrcode==8.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
django-widget-tweaks==1.5.0
django-crispy-forms==2.3
crispy-bootstrap5==2.0.0
dj-database-url==2.3.0